pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "concept-modelling"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["torch", "faiss-cpu", "numpy"]
